Better globbing for fuzzy search
The (excellent) fuzzy search feature supports an option titled "Directories to omit from matching".

However, the parameters are assumed to appear only in the beginning of the path. For instance, `node_modules` is correctly ignored, but only the `node_modules` at the _root_ of your project. If you have multiple projects in your Git repo, each with its own `node_modules`, these are not excluded.

This PR tweaks the arguments to `glob` such that the ignore patterns are passed separately and can match any subdirectory of your project. This way, both `/myproject/node_modules` and `/myproject/mysubproject/node_modules` are correctly ignored.

Additionally, the unnecessary extra globbing against the project root was removed - it turns out `**` can resolve to the project root already - the problem was that you typically weren't seeing file matches from the project root because `glob` returns those at the end of its (potentially very long) array of files that it found, so they were "falling off" the list of the top six matches.

To address this (for now), I simply `reverse()`d the glob array such that the root files appear at the beginning where you would expect to see them.